Add files via upload
This script used yougetsignal dot com to perform passive recon by using the site to perform a port scan on our behalf. This is an excersise in learning Python web requests. I'll come back and improve this over time as I learn more Python. 
To do: Add the ability to specify a range of ports other than the default, or scan all ports.